Revert "remove explicit world age propagation; turns out tests pass regardless as long as we reflect on world=typemax(UInt)"
This reverts commit be7f8d3aef54690152f006ff8bc5eb0e787f0005.

It turns out that the only reason tests were passing after the reverted commit was because Cassette’s new API (which the tests use) generates a closure for each `@overdub` invocation. Thus, recompilation was triggering on each new closure type for each test, and none of the tests were stressing world-age based recompilation.